pyproject.toml
[project]
name = "model-armor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-modelarmor>=0.1.0",
    "python-dotenv>=1.0.1",
]
